Project/requirements.txt
flask
flask-login
flask-bcrypt
flask-sqlalchemy
python-dotenv
requests
